ComSense Pt 2
And lastly for now, anyone is welcome to incorporate ComSenMod into
their own works, but I'd appreciate it if you'd credit me, Crater
Creator, if you do. Or to put it formally, ComSenMod is released under
the CC BY license. I will gladly take an opportunity to put out a Terms
of Use that isn't 10 pages of legalese